compton: put the settings to be changed in a delayed queue.
When a Settings.changed event is fired, the modified setting is now stored in an HashTable instead of
directly set it into the ComptonConfiguration (with the subsequent dump).

A 700ms timeout triggers the actual setting application. Said timeout is reset whenever a new Settings.changed event
is fired and the setting application step has not been executed yet.

While there is a somewhat big disadvantage with this change (the user feedback is delayed), the compton
configuration file is dumped only one time per cycle and the delay makes the usage of UI sliders easier.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>